Semanas/Semana3/Dia6/requirements.txt
# Dependências do Dia 6 - Testes Automatizados
# Baseado no Dia 5, com adições para testes

# FastAPI e servidor
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Autenticação e segurança
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Rate limiting
slowapi>=0.1.9

# LangChain e OpenAI
langchain>=0.1.0
langchain-openai>=0.0.5

# Validação
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Variáveis de ambiente
python-dotenv>=1.0.0

# Testes
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
